+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [21.1.16] - 2021-05-20
+### Added
+- Wipe custom object during product reset
+
+### Changed
+- Removed checkpoint object from syncing API calls to keep error logs clean
+
 ## [21.1.15] - 2021-05-18
 [fbbe07ba4b2ae56c4a8c5f44163cb68ed0e5f3c0](https://github.com/stickyio/link_stickyio_preview/commit/fbbe07ba4b2ae56c4a8c5f44163cb68ed0e5f3c0)
 ### Changed